ref: rename CampaignProcessor message handler classes and update namespaces
